Finalized Elevator
The implementations for the...
+TalonFX
+SparkMax
+Simulation
Have been finalized. To run this you would only need to modify constants to make it specific to your robot.

Because there was no acceleration methods for the sparkmax or sim, I have made one from scratch using the timer class from java.lang. It takes velocity samples over a set periodic of time and returns the differences between periods divided by the sample rate.